requirements.txt
# Core
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Faker==20.1.0

# Compiler & NLP
nltk==3.8.1
regex==2023.10.3

# Automata
transitions==0.9.0

# AI & LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
transformers==4.36.2
torch==2.1.2

# Dashboard
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
folium==0.15.1
streamlit-folium==0.17.0
altair==5.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utils
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3
tabulate==0.9.0
rich==13.7.0
click==8.1.7
